refactor: separate torrent repository trait from implementations
There are now more implementations.
