+++ packages/torrent-repository/src/lib.rs
@@ -1,5 +1,7 @@
 use std::sync::Arc;
 
+use repository::rw_lock_std::RwLockStd;
+use repository::rw_lock_tokio::RwLockTokio;
 use repository::skip_map_mutex_std::CrossbeamSkipList;
 use torrust_tracker_clock::clock;
 
@@ -10,12 +12,12 @@ pub type EntrySingle = entry::Torrent;
 pub type EntryMutexStd = Arc<std::sync::Mutex<entry::Torrent>>;
 pub type EntryMutexTokio = Arc<tokio::sync::Mutex<entry::Torrent>>;
 
-pub type TorrentsRwLockStd = repository::RwLockStd<EntrySingle>;
-pub type TorrentsRwLockStdMutexStd = repository::RwLockStd<EntryMutexStd>;
-pub type TorrentsRwLockStdMutexTokio = repository::RwLockStd<EntryMutexTokio>;
-pub type TorrentsRwLockTokio = repository::RwLockTokio<EntrySingle>;
-pub type TorrentsRwLockTokioMutexStd = repository::RwLockTokio<EntryMutexStd>;
-pub type TorrentsRwLockTokioMutexTokio = repository::RwLockTokio<EntryMutexTokio>;
+pub type TorrentsRwLockStd = RwLockStd<EntrySingle>;
+pub type TorrentsRwLockStdMutexStd = RwLockStd<EntryMutexStd>;
+pub type TorrentsRwLockStdMutexTokio = RwLockStd<EntryMutexTokio>;
+pub type TorrentsRwLockTokio = RwLockTokio<EntrySingle>;
+pub type TorrentsRwLockTokioMutexStd = RwLockTokio<EntryMutexStd>;
+pub type TorrentsRwLockTokioMutexTokio = RwLockTokio<EntryMutexTokio>;
 
 pub type TorrentsSkipMapMutexStd = CrossbeamSkipList<EntryMutexStd>;
